eval_code/eval_rg/requirements.txt
lightning
transformers
peft
huggingface_hub
tqdm
torch
numpy
pillow
nltk
rouge-score
bert-score
